Refactoring for eliminate utils package #47
