+++ analyzer/analyzer.go
@@ -2,6 +2,7 @@ package analyzer
 
 import (
 	"fmt"
+	"github.com/rodrigo-brito/gocity/lib/file"
 	"go/ast"
 	"go/parser"
 	"go/token"
@@ -10,8 +11,6 @@ import (
 	"strings"
 
 	log "github.com/sirupsen/logrus"
-
-	"github.com/rodrigo-brito/gocity/utils"
 
 	"github.com/rodrigo-brito/gocity/lib"
 )
@@ -71,7 +70,7 @@ func (a *analyzer) Analyze() (map[string]*NodeInfo, error) {
 		}
 
 		fileSet := token.NewFileSet()
-		if f.IsDir() || !utils.IsGoFile(f.Name()) || a.IsInvalidPath(path) {
+		if f.IsDir() || !file.IsGoFile(f.Name()) || a.IsInvalidPath(path) {
 			return nil
 		}
