+++ analyzer/visitor.go
@@ -1,10 +1,9 @@
 package analyzer
 
 import (
+	"github.com/rodrigo-brito/gocity/lib/file"
 	"go/ast"
 	"go/token"
-
-	"github.com/rodrigo-brito/gocity/utils"
 )
 
 type NodeInfo struct {
@@ -34,7 +33,7 @@ func (v *Visitor) Visit(node ast.Node) ast.Visitor {
 
 	switch definition := node.(type) {
 	case *ast.ValueSpec: // Atributes
-		identifier := utils.GetIdentifier(v.Path, v.PackageName, "")
+		identifier := file.GetIdentifier(v.Path, v.PackageName, "")
 
 		if _, ok := v.StructInfo[identifier]; !ok {
 			v.StructInfo[identifier] = new(NodeInfo)
@@ -45,7 +44,7 @@ func (v *Visitor) Visit(node ast.Node) ast.Visitor {
 
 	case *ast.TypeSpec: // Structs
 		if structObj, ok := definition.Type.(*ast.StructType); ok {
-			identifier := utils.GetIdentifier(v.Path, v.PackageName, definition.Name.Name)
+			identifier := file.GetIdentifier(v.Path, v.PackageName, definition.Name.Name)
 
 			if _, ok := v.StructInfo[identifier]; !ok {
 				v.StructInfo[identifier] = new(NodeInfo)
@@ -71,7 +70,7 @@ func (v *Visitor) Visit(node ast.Node) ast.Visitor {
 			}
 		}
 
-		identifier := utils.GetIdentifier(v.Path, v.PackageName, structName)
+		identifier := file.GetIdentifier(v.Path, v.PackageName, structName)
 
 		if _, ok := v.StructInfo[identifier]; !ok {
 			v.StructInfo[identifier] = new(NodeInfo)
