+++ lib/file/utils.go
@@ -1,4 +1,4 @@
-package utils
+package file
 
 import (
 	"fmt"
