+++ model/node.go
@@ -2,10 +2,9 @@ package model
 
 import (
 	"fmt"
+	"github.com/rodrigo-brito/gocity/lib/file"
 	"sort"
 	"strings"
-
-	"github.com/rodrigo-brito/gocity/utils"
 
 	"github.com/rodrigo-brito/gocity/analyzer"
 )
@@ -104,7 +103,7 @@ func (n *Node) GenerateChildrenPosition() {
 func getPathAndFile(fullPath string) (paths []string, fileName, structName string) {
 	pathlist := strings.Split(fullPath, "/")
 	paths = pathlist[:len(pathlist)-1]
-	fileName, structName = utils.GetFileAndStruct(pathlist[len(pathlist)-1])
+	fileName, structName = file.GetFileAndStruct(pathlist[len(pathlist)-1])
 	return
 }
